Button on_click and Menu changes
on_click functions now take a single parameter. When called the button
that received the click is passed as the value of that parameter.

Add alert() to dialog.py to make showing alert dialogs simpler.

Mouse events that occur off a menu when a menu is showing now dismiss
the menu.

ESC quits a popup menu.